+++ src/main/java/org/opensearch/knn/index/codec/nativeindex/remote/RemoteIndexBuildStrategy.java
@@ -6,6 +6,7 @@
 package org.opensearch.knn.index.codec.nativeindex.remote;
 
 import com.google.common.annotations.VisibleForTesting;
+import lombok.NonNull;
 import lombok.extern.log4j.Log4j2;
 import org.opensearch.cluster.ClusterName;
 import org.opensearch.cluster.metadata.RepositoryMetadata;
@@ -20,7 +21,7 @@ import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.index.codec.nativeindex.NativeIndexBuildStrategy;
 import org.opensearch.knn.index.codec.nativeindex.model.BuildIndexParams;
 import org.opensearch.knn.index.engine.KNNLibraryIndexingContext;
-import org.opensearch.knn.index.engine.faiss.FaissHNSWMethod;
+import org.opensearch.knn.index.engine.ResolvedIndexSpec;
 import org.opensearch.knn.index.remote.RemoteIndexWaiter;
 import org.opensearch.knn.index.remote.RemoteIndexWaiterFactory;
 import org.opensearch.knn.index.vectorvalues.KNNVectorValues;
@@ -235,7 +236,8 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
                 indexInfo,
                 repositoryContext.blobStoreRepository.getMetadata(),
                 repositoryContext.blobPath.buildAsString() + repositoryContext.blobName,
-                knnLibraryIndexingContext.getLibraryParameters()
+                knnLibraryIndexingContext.getLibraryParameters(),
+                knnLibraryIndexingContext.getResolvedSpec()
             );
             remoteBuildResponse = client.submitVectorBuild(buildRequest);
             success = true;
@@ -337,14 +339,14 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
         return new RepositoryContext(repository, blobPath, vectorRepositoryAccessor, blobName);
     }
 
-    private static String determineVectorDataType(final VectorDataType dataType, final Map<String, Object> parameters) {
+    private static String determineVectorDataType(final VectorDataType dataType, @NonNull final ResolvedIndexSpec resolvedSpec) {
         if (dataType == VectorDataType.FLOAT) {
             // SQ 1-bit sends fp32 vectors. Once support is added for building 1 bit SQ graphs
             // in the Remote Vector Index Builder, then this can be modified to send 1 bit SQ vectors.
-            if (FaissHNSWMethod.isSQOneBitIndex(dataType, parameters)) {
+            if (resolvedSpec.isFaissSQOneBit()) {
                 return dataType.getValue();
             }
-            if (FaissHNSWMethod.isFloat16Index(dataType, parameters)) {
+            if (resolvedSpec.isFP16QuantizedIndex()) {
                 return FLOAT16_VECTOR_TYPE_STRING;
             }
         }
@@ -357,8 +359,8 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
      * When true, the RVIB writes only the HNSW graph, and the data node stitches it with locally-stored
      * flat vectors at search time via {@link org.opensearch.knn.memoryoptsearch.faiss.FaissFlatIndexFactory}.
      */
-    private static boolean shouldSkipStoredVectors(final VectorDataType vectorDataType, final Map<String, Object> parameters) {
-        return FaissHNSWMethod.isSQOneBitIndex(vectorDataType, parameters);
+    private static boolean shouldSkipStoredVectors(final VectorDataType vectorDataType, @NonNull final ResolvedIndexSpec resolvedSpec) {
+        return resolvedSpec.isFaissSQOneBit();
     }
 
     /**
@@ -369,6 +371,7 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
      * @param repositoryMetadata RepositoryMetadata object
      * @param fullPath           Full blob path + file name representing location of the vectors/doc IDs (excludes repository-specific prefix)
      * @param parameters         Map of parameters to be parsed and passed to the remote build service
+     * @param resolvedSpec       ResolvedIndexSpec for spec-based decisions
      * @throws IOException if an I/O error occurs
      */
     static RemoteBuildRequest buildRemoteBuildRequest(
@@ -376,7 +379,8 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
         BuildIndexParams indexInfo,
         RepositoryMetadata repositoryMetadata,
         String fullPath,
-        Map<String, Object> parameters
+        Map<String, Object> parameters,
+        @NonNull ResolvedIndexSpec resolvedSpec
     ) throws IOException {
         final String repositoryType = repositoryMetadata.type();
         final String containerName;
@@ -387,7 +391,7 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
             );
         }
 
-        final String vectorDataType = determineVectorDataType(indexInfo.getVectorDataType(), parameters);
+        final String vectorDataType = determineVectorDataType(indexInfo.getVectorDataType(), resolvedSpec);
 
         KNNVectorValues<?> vectorValues = decorateVectorValuesSupplier(indexInfo).get();
         initializeVectorValues(vectorValues);
@@ -404,7 +408,7 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
             .vectorDataType(vectorDataType)
             .engine(indexInfo.getKnnEngine().getName())
             .indexParameters(indexInfo.getKnnEngine().createRemoteIndexingParameters(parameters))
-            .skipStoredVectors(shouldSkipStoredVectors(indexInfo.getVectorDataType(), parameters))
+            .skipStoredVectors(shouldSkipStoredVectors(indexInfo.getVectorDataType(), resolvedSpec))
             .build();
     }
 }